add ability to specify start_block to a dataset (#21)
